Default queueOptions[name] to null
A producer doesn't need to specify any queue options. If you don't specify queue options the default value for `queueOptions[name]` is currently an empty string. Auto setup fabric then runs queueDeclare on the producer which checks whether `queueOptions[name] !== null` to see if it needs to set up a queue. This fails with empty string and a queue with a blank name - resulting in an automatically created name by rabbitmq - is created. This ends up creating tons of unwanted queues if you keep it running for a while.

So the fix is simple, default to null for the queue name.