Use 2.12.2
This version is the only one that disallows installation on PHP 8.0. While 2.12.0 and 2.12.1 are installable on php: >=5.6.3.

In SimpleBus we test our dependencies with `lowest` and therefore this fails on PHP 8.0. 

See PR with failing tests https://github.com/SimpleBus/SimpleBus/pull/115

Of course, we could add a conflict, but I believe this should be done here.